Remove sharing lock from cache mount on RUN
The cache was empty when using a `locked` sharing type.
Reverting to the default sharing of `shared` as the build should
not be affected with the current workflow.

https://github.com/moby/buildkit/blob/f2a6e83adcb0295099870489b76d3ce74d6f7f42/frontend/dockerfile/docs/syntax.md#run---mounttypecache

> This mount type allows the build container to cache directories for
compilers and package managers.

|Option               |Description|
|---------------------|-----------|
|`sharing`            | One of `shared`, `private`, or `locked`.
Defaults to `shared`. A `shared` cache mount can be used concurrently by
multiple writers. `private` creates a new mount if there are multiple
writers. `locked` pauses the second writer until the first one releases
the mount.|